wip: unit-tests: Check installed IPsec SAs in child-rekey tests
wip: The behavior during rekey collisions is not ideal yet (we might
install an SA just do delete it).
Ideally we'd also want to delay deinstallation of the rekeyed inbound
SA (this might require a method to deinstall only the outbound SA and a
new state to track it). The delay might be tricky to test as we currently
don't use the ike_sa_manager that the jobs use (and we might need a
mock scheduler/processor).